# eyeson-go-server/static/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "EyesOn SIM Management API",
...
    "version": "1.0.0",
    "contact": {
      "name": "EyesOn API Support",
      "email": "api-support@eyeson.local",
      "url": "https://github.com/alexgavs/eyeson-go"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    },
    "x-logo": {
      "url": "https://example.com/logo.png",
      "altText": "EyesOn API"
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local Development Server"
    },
    {
      "url": "http://127.0.0.1:5000",
      "description": "Local Server (IP Address)"
    },
    {
      "url": "https://api.eyeson.local",
      "description": "Production Server (Example)"
    }
  ],
...
  "paths": {
...
    "/api/v1/api-status": {
      "get": {
        "tags": ["Statistics"],
        "summary": "Check upstream API connection status",
        "description": "Performs a health check on the upstream Pelephone EyesOnT API connection and returns detailed connection information including API URL, credentials (masked), and current SIM count.\n\n**Required Role**: Administrator only\n\n**Note**: This endpoint makes a real API call to verify connectivity.",
        "operationId": "getApiStatus",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {
            "description": "API status retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStatusResponse"
                },
                "examples": {
                  "connected": {
                    "summary": "Connected",
                    "description": "Successful connection to upstream API",
                    "value": {
                      "status": "connected",
                      "details": {
                        "api_url": "https://eot-portal.pelephone.co.il:8888",
                        "api_user": "apiuser",
                        "total_sims": "50",
                        "cache_age": "127"
                      }
                    }
                  },
                  "error": {
                    "summary": "Connection Error",
                    "description": "Failed to connect to upstream API",
                    "value": {
                      "status": "error",
                      "error": "Connection timeout after 30 seconds"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Valid authentication token required"
          },
          "403": {
            "description": "Forbidden - Administrator role required"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JSON Web Token authentication. Obtain a token via POST /api/v1/auth/login and include it in the Authorization header."
      }
    },
    "schemas": {
...
      "ApiStatusResponse": {
        "type": "object",
        "description": "Upstream API connection status",
        "properties": {
          "status": {
            "type": "string",
            "description": "Connection status",
            "enum": ["connected", "error"],
            "example": "connected"
          },
          "details": {
            "type": "object",
            "description": "Connection details (present when status is connected)",
            "properties": {
              "api_url": {
                "type": "string",
                "description": "Upstream API base URL",
                "example": "https://eot-portal.pelephone.co.il:8888"
              },
              "api_user": {
                "type": "string",
                "description": "API authentication username",
                "example": "apiuser"
              },
              "total_sims": {
                "type": "string",
                "description": "Total SIM count from upstream API",
                "example": "50"
              },
              "cache_age": {
                "type": "string",
                "description": "Cache age in seconds",
                "example": "127"
              }
            }
          },
          "error": {
            "type": "string",
            "description": "Error message (present when status is error)",
            "example": "Connection timeout"
          }
        }
      },
...
    }
  }
}